backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.11.0
websockets>=13.0
python-multipart==0.0.6
sentence-transformers==2.7.0
chromadb==0.4.22
PyPDF2==3.0.1
python-dotenv==1.0.0
supabase
